+++ src/docs/components/markdown/pre.svelte
@@ -21,6 +21,7 @@
 </script>
 
 {#if show}
+	<!-- svelte-ignore a11y-no-noninteractive-tabindex -- This is needed to be acessible -->
 	<pre
 		use:setCodeString
 		use:setIsPPBlock
@@ -29,6 +30,7 @@
 			isPPBlock === undefined && 'data-[non-pp]:hidden',
 			className
 		)}
+		tabindex="0"
 		{...$$restProps}>
     <slot />
 	</pre>
@@ -51,3 +53,10 @@
 		</button>
 	{/if}
 {/if}
+
+<style>
+	/* Override theme colors for WCAG concerns */
+	pre :global([style*='color: #6A737D']) {
+		color: #727e8b !important;
+	}
+</style>