+++ src/docs/previews/accordion/disabled/css.svelte
@@ -34,11 +34,7 @@
 	{#each items as { id, title, description, disabled }, i}
 		<div melt={$item(id)} class="item">
 			<h2>
-				<button
-					id={i === 0 ? 'accordion-trigger' : undefined}
-					melt={$trigger({ value: id, disabled })}
-					class="trigger"
-				>
+				<button melt={$trigger({ value: id, disabled })} class="trigger">
 					{title}
 				</button>
 			</h2>
@@ -51,7 +47,7 @@
 	{/each}
 </div>
 
-<style lang="postcss">
+<style>
 	.root {
 		margin-left: auto;
 		margin-right: auto;