+++ src/docs/previews/avatar/main/css.svelte
@@ -40,7 +40,7 @@
 	</div>
 </div>
 
-<style lang="postcss">
+<style>
 	.root {
 		display: flex;
 		width: 100%;