+++ src/docs/previews/collapsible/main/tailwind.svelte
@@ -19,6 +19,7 @@
 			class="relative h-6 w-6 place-items-center rounded-full bg-white text-sm
             text-magnum-700 shadow-lg hover:opacity-75 data-[disabled]:cursor-not-allowed
             data-[disabled]:opacity-75"
+			aria-label="Toggle"
 		>
 			<div class="abs-center">
 				{#if $open}